# public/api_doc/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/employee": {
            "post": {
                "tags": [
                    "No authorization needed"
                ],
                "summary": "Creación de un empleado nuevo",
                "operationId": "a1ead0b448ceb79d81a5701d2fda67a7",
                "requestBody": {
                    "description": "Información del nuevo empleado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "surname",
                                    "phone",
                                    "birthDate"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "birthDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "datetime"
                                    },
                                    "creationDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "datetime"
                                    },
                                    "lastConnection": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "datetime"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "profileImage": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "marketingMails": {
                                        "type": "boolean",
                                        "example": "boolean"
                                    },
                                    "centers": {
                                        "type": "relation",
                                        "example": "relation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Alguno de los parámetros no es válido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Datos no válidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "El user ha sido creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "roles": {
                                                "type": "json",
                                                "example": "json"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "gender": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "birthDate": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "datetime"
                                            },
                                            "creationDate": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "datetime"
                                            },
                                            "lastConnection": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "datetime"
                                            },
                                            "dni": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "profileImage": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "marketingMails": {
                                                "type": "boolean",
                                                "example": "boolean"
                                            },
                                            "centers": {
                                                "type": "relation",
                                                "example": "relation"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": {
                                    "message": "El user ha sido editado correctamente",
                                    "data": {
                                        "id": "3",
                                        "email": "string",
                                        "roles": "json",
                                        "password": "string",
                                        "name": "string",
                                        "surname": "string",
                                        "phone": "string",
                                        "gender": "string",
                                        "birthDate": "datetime",
                                        "creationDate": "datetime",
                                        "lastConnection": "datetime",
                                        "dni": "string",
                                        "profileImage": "string",
                                        "marketingMails": "boolean",
                                        "center": "relation"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/cambiarPassword": {
            "post": {
                "tags": [
                    "No authorization needed"
                ],
                "summary": "Cambio de contraseña de una cuenta",
                "operationId": "e85bc8073946abf3ede2cd17bd254a60",
                "requestBody": {
                    "description": "Token y nueva contraseña",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recover_token",
                                    "new_password"
                                ],
                                "properties": {
                                    "recover_token": {
                                        "type": "string",
                                        "example": "ZHjyt79MliaQ7lNeMbi4k32MvHcyut+ypN0BDsO+BNXO5nUb+A+f/JYur8KKBrO4%7C3kAmrA60ALEMKIAGjKAveQ=="
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3 (hasheado en SHA256)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Alguna de las propiedades del request body no es válida o bien el token ya ha caducado o ha sido usado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Datos no válidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "El user ha cambiado su contraseña con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password cambiado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/refreshToken": {
            "post": {
                "tags": [
                    "No authorization needed"
                ],
                "summary": "Actualizar token de seguridad",
                "operationId": "52368e358857fd8844e8680ae6121904",
                "responses": {
                    "200": {
                        "description": "El token ha sido actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "refresh_token": {
                                            "type": "string",
                                            "example": "f731efc553108b8940b8028aa2f9546ba052d4399efcf835ccf3d270a35000c073b88df133f108a0af1a4799f5bf0614df845b0b76f1b07dd21b54f4cd366383"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}